updated alignment_counter, removed alignment_counter2
