+++ gffquant/counters/__init__.py
@@ -3,7 +3,6 @@
 
 """module docstring"""
 
-from .alignment_counter2 import AlignmentCounter
+from .alignment_counter import AlignmentCounter
 from .region_counter import RegionCounter
-from .seq_counter import UniqueSeqCounter, AmbiguousSeqCounter
 from .count_manager import CountManager
